% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supFunctions.R
\name{data_gen}
\alias{data_gen}
\title{data_gen_gauss}
\usage{
data_gen(n, chpts = NULL, means = 0, noise = 1, type = "gauss")
}
\arguments{
\item{n}{number of data points}

\item{chpts}{a vector of increasing change-point indices. Last index is always less than 'n'.
By default, 'chpts = NULL' (the data without change-points).}

\item{means}{vector of successive means for data, by default 'means = 0'.}

\item{noise}{standard deviation of an additional normal noise, by default 'noise = 1'.}
}
\value{
a vector of size n generated data with a given values of means by the segmentation.
}
\description{
Generation of gaussian data with a given values of means and changepoints
}
\examples{
data <- data_gen_gauss(n = 100, chpts = 50, means = c(1,5), noise = 1)
}
