% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastFPOPmidVersion.R
\name{fastFPOPmidVersion}
\alias{fastFPOPmidVersion}
\title{fastFPOPmidVersion}
\usage{
fastFPOPmidVersion(data, beta, affiche = FALSE)
}
\arguments{
\item{data}{is a vector of data points (a univariate time series)}

\item{beta}{is a value of penalty (a non-negative real number)}

\item{affiche}{if it is TRUE Table v and the plots of functions q^{i}_{t+1} are displayed on the screen at each iteration}

\item{cost}{is a number equal to the global cost}
}
\value{
a vector of changepoints, global cost
}
\description{
Functional Pruned Optimal Partitioning 1D Algorithm ( fast version with Gaussian loss)
}
\examples{
library(fastFPOP)
data <- data_gen(n = 10, chpts = 5, means = c(0,5), noise = 1)
fastFPOPmidVersion(data, 2*log(10),  affiche = TRUE)
fastFPOPmidVersion(data, 2*log(100),  affiche = FALSE)
}
