% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supFunctions.R
\name{plot_V}
\alias{plot_V}
\title{plot_V in the time t+1}
\usage{
plot_V(data, V, mi, beta, t)
}
\arguments{
\item{data}{is a vector of data points (a univariate time series)}

\item{V}{is a table of table of triplets and quadruplets}

\item{mi}{is the vector of the min values}

\item{beta}{is a value of penalty}

\item{t}{is the value of t}
}
\value{
the plots of the cost functions contained in the table V
}
\description{
Plots of the cost functions contained in the table V
}
\examples{

}
