% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supFunctions.R
\name{q_abp1_theta}
\alias{q_abp1_theta}
\title{q_abp1_theta}
\usage{
q_abp1_theta(a, b, theta, csd_a, csd2_a, csd_b, csd2_b, beta, m_am1)
}
\arguments{
\item{a}{is a value of i}

\item{b}{is a value of t}

\item{theta}{is a function argument}

\item{csd_a}{is a value of cumsum value data(1:i)}

\item{csd2_a}{is a value of cumsum value data^2(1:i)}

\item{csd_b}{is a value of cumsum value data(1:t+1)}

\item{csd2_b}{is a value of cumsum value data^2(1:t+1)}

\item{beta}{is a value of penalty}

\item{m_am1}{is the value of m_{i-1}}
}
\value{
a value of q^i_{t+1}-function  in theta
}
\description{
Value of q^i_{t+1}(theta)-function  in theta( using cumsum of data)
}
