% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supFunctions.R
\name{qin}
\alias{qin}
\title{qin}
\usage{
qin(theta, data, min, beta, i, n)
}
\arguments{
\item{theta}{is a function argument}

\item{data}{is a vector of data points (a univariate time series)}

\item{min}{is the vector of the min values}

\item{beta}{is a value of penalty}

\item{i}{is the left border value}

\item{n}{is the right border value}
}
\value{
a value of q^i_{n}-function in theta
}
\description{
Value of q^i_{n}-function  in theta
}
\examples{

}
